ssdsdd

+++ src/app/services/user.service.ts
@@ -1,10 +1,9 @@
-import { Injectable } from '@angular/core';
+import { Injectable, EventEmitter } from '@angular/core';
 import { AngularFirestore,AngularFirestoreCollection,AngularFirestoreDocument } from 'angularfire2/firestore';
 import { Observable, from, of } from 'rxjs';
 import { map, filter, catchError, mergeMap } from 'rxjs/operators';
 import { User } from '../models/user';
 import { Router, CanActivate } from '@angular/router';
-import { EventEmitter } from 'events';
 import { MessageService } from './message.service';
 
 
@@ -12,16 +11,20 @@ import { MessageService } from './message.service';
   providedIn: 'root'
 })
 export class UserService {
-  usersCollectionRef: AngularFirestoreCollection<User>;
+  private usersCollectionRef: AngularFirestoreCollection<User>;
   users$: Observable<User[]>;
   currentUser$:Observable<User>;
-  collectionName:string;
+
+  private selectedClient:User;
+  private collectionName:string;
+  private phonePrefix:string;
 
   private userDoc: AngularFirestoreDocument<User>;
   
 
   constructor(private afs: AngularFirestore, private router: Router,private msgService:MessageService) {
     this.collectionName='users';
+
     this.usersCollectionRef = this.afs.collection(this.collectionName);
     this.users$ = this.usersCollectionRef.valueChanges();
     
@@ -30,39 +33,53 @@ export class UserService {
       {
           return actions.map(action=>
           {
-            const data = action.payload.doc.data() as User;
+            const data = action.payload.doc.data();
             const id = action.payload.doc.id;
             return { id, ...data };
           })   
       }));
    }
+  setPhonePrefix(phonePrefix:string):void{
+    if (phonePrefix) {
+      this.phonePrefix=phonePrefix;
+    }
+  }
+  onSelectedClient(client:User){
+     if (client) {
+      this.selectedClient=client;
+      this.router.navigate(['user']);
+     }
+  }
+  getSelectedClient():User{
+     return this.selectedClient;
+  }
 
   addUser(user:User ):void
   {
     if(user)
     {
-     let res= this.getUser(user.email).then(res=>{
+      let res= this.getUser(user.email).then(res=>{
         if(!res)
         {
-           this.usersCollectionRef.doc(user.email).set({firstName:user.firstName,
-                                                        lastName:user.lastName,
-                                                        email:user.email,
-                                                        phone:user.phone,
-                                                        admin:false})
+          user.admin=false;
+          const userForDb=<User>user;
+           this.usersCollectionRef.doc(userForDb.email).set(userForDb)
            .then((complit)=>this.router.navigate(['home']),
                   (error)=>console.log(error));
         }
         else{
           this.msgService.sendMessage("alert-warning","this email is allready registered");
-        }});
-        
-    }
-   }
+        }});    
+    }}
+
+
+  getListOfUsers():Observable<any>{
+    return this.usersCollectionRef
+    .get();
+  }
 
    getUser(email:string):Promise<any>{
 
-    //  this.userDoc = this.afs.doc<User>('users/'+email);
-    //  return this.userDoc.valueChanges();
       return this.usersCollectionRef
              .doc(email)
              .ref
@@ -75,11 +92,24 @@ export class UserService {
    
 
     updateUser(user: User) {
-      this.usersCollectionRef.doc(user.id).update(user);
+      if (user) {
+        this.usersCollectionRef.doc(user.id).update(user);
+      }
+      
     }
     
     deleteUser(user: User) {
-      this.usersCollectionRef.doc(user.id).delete();
+      if (user) {
+        this.usersCollectionRef.doc(user.id).delete();
+      }
+      
     }
   
+     getPnoneNumber(phone:string){
+      let fullNumber=this.phonePrefix + phone;
+     return this.usersCollectionRef
+        .ref
+        .where("phone", "==", fullNumber)
+        .get();
+    }
 }
